# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare AI Billing API (example)
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/subscriptions/create:
    post:
      summary: Create a subscription and grant initial credits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '200':
          description: Subscription created
  /api/v1/accounts/{accountId}/summary:
    get:
      summary: Get account credit summary (for dashboard)
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id: { type: string }
                  available_credits: { type: integer }
                  reserved_credits: { type: integer }
                  lifetime_earned: { type: integer }
                  lifetime_spent: { type: integer }
  /webhook/stripe:
    post:
      summary: Stripe webhook endpoint (verify signature)
      requestBody:
        required: true
        content:
          application/json: {}
      responses:
        '200':
          description: OK
components:
  schemas:
...
    CreateSubscriptionRequest:
      type: object
      properties:
        user_id: { type: string }
        organization_id: { type: string }
        plan_code: { type: string }